+++ src/main/groovy/net/thesilkminer/gradle/plugin/languagechecker/LineTemplate.groovy
@@ -1,0 +1,6 @@
+package net.thesilkminer.gradle.plugin.languagechecker
+
+interface LineTemplate {
+
+    String fill(Map<String, String> translations)
+}